src/aasist/pipeline/components/download_dataset/runtime-requirements.txt
# Generated by KFP.
python-dotenv==1.1.1
requests==2.32.4
tqdm==4.67.1